Extract create value for Color to helper for:
- PredefinedColorParser
